xci: Configure ansible verbosity based on job type
Daily jobs will use pinned versions of bifrost and openstack-ansible in 
order to make sure we use known good versions for them since focus of
the daily jobs is  OPNFV platform testing.

Ansible log can easily become massive depending on the level it is set 
and it is not so relevant for daily jobs to have it like this due to
less probability of dailies failing to provision/deploy.

On the other hand, the jobs that verify/test bifrost and openstack-ansible 
will have max verbosity enabled to troubleshoot these components themselves.

New set of jobs will be created running periodically against the HEAD
of master for bifrost and openstack-ansible to catch working versions
and help us promote them to use for daily platform jobs. Verbosity of these
jobs will be enabled like how we are doing for bifrost-verify jobs.

Change-Id: Ic9fd2ff418fdaa3ad7fbfc22d8ec2dc1174df65f
Signed-off-by: Fatih Degirmenci <fatih.degirmenci@ericsson.com>