xci: Add ansible files for sandbox noha flavor
This patch adds noha target host configuration playbook, inventory, variable
files and so on to use until we have time to fix things more properly and some
other functionality becomes available in upstream. This approach will result in
duplicates but it is something we can perhaps live with.

NFS configuration is converted into a role and the rest is adjusted accordingly.

The playbook to configure target hosts will be same with the one we will
have for ha flavor but for different for the flavor mini.

The user variables and openstack user configuration will be different between
flavors and keeping deployment specific configuration.

Change-Id: I8faceae846285fc3e4260cbced5b740a9b0c3766
Signed-off-by: Fatih Degirmenci <fatih.degirmenci@ericsson.com>