xci: Add ansible files for sandbox ha flavor
This patch adds noha target host configuration playbook, inventory, variable
files and so on to use until we have time to fix things more properly and some
other functionality becomes available in upstream. This approach will result in
duplicates but it is something we can perhaps live with.

The playbook to configure target hosts will be same with the one we will
have for noha flavor but for different for the flavor mini.

The user variables and openstack user configuration will be different between
flavors and keeping deployment specific configuration.

Change-Id: I4e2b375b9f8f6bd5f8c5da91a522b78d61a58125
Signed-off-by: Fatih Degirmenci <fatih.degirmenci@ericsson.com>