xci: Fix the haproxy service and gnocchi db sync problem
This patch reflects the fix Tianwei found for the keystone and gnocchi
issues with HA deployment.

https://gerrit.opnfv.org/gerrit/#/c/32367/

The keystone issue should be fixed upstream but until that happens, we
can use the patch here to move us forward.

Previously removed OpenStack services have also been enabled to
try as close deployment as possible to upstream openstack-ansible.

Change-Id: Ib7b1cfe0f363381466b43a3249667c638c09d9fd
Signed-off-by: Fatih Degirmenci <fatih.degirmenci@ericsson.com>