prototypes: xci: xci-deploy.sh: Only allow XCI on Ubuntu
There is no support for openSUSE or CentOS right now yet the
script runs happily on these hosts until it breaks in horrible
ways. The only place we explicitly ask for Ubuntu hosts is the wiki
which can easily be missed by newcomers. As such, modify the script
to only run on Ubuntu hosts.

Change-Id: If9b355da51367fc95948a2b12a7f6c4d6b3bbbbf
Signed-off-by: Markos Chandras <mchandras@suse.de>