xci: Bump OSA SHA to Pre-Pike
This change bumps osa sha to 4d39f2cc. [1] The reason for going to a later
osa sha is to include ODL and Tacker commits in upstream to ease the work
done for ODL and SFC scenarios. However, this osa sha seems to be having
issues with the flavor ha so the users should be aware of the problems
which is added into the readme as known issue and the flavor is disabled
in xci-deploy.sh.

Another reason to bump sha right now is that upstream changed a lot during
last 2 months and waiting longer will probably result in more changes, making
it harder to move to a later osa sha even harder.

Other changes included by this change are

- update ansible role requirements to move to shas pinned by upstream for pike.
- add odl and tacker roles into ansible role requirements.
- update how aio is deployed due to removal of an upstream script. (this could
perhaps have been done in a better way but the time pressure required this to
be fixed in the way as proposed.)
- ensure facts are gathered for all the hosts and containers before
running setup-infrastructure.yml playbook.
- stop grepping for failed in the log of setup-hosts.yml playbook due
to existence of rescue block in rabbitmq_cluster_create.yml [2].
- reduce the resource needs by updating the user_variables that are
based on upstream osa gate which uses aio. The values for none-aio xci
flavors will be adjusted in a separate change once the sha bump is complete.
- fix a bug in ha-vars.

[1] https://review.openstack.org/gitweb?p=openstack/openstack-ansible.git;a=commit;h=4d39f2cc29417153780210fc0bb86223387e9968
[2] https://review.openstack.org/gitweb?p=openstack/openstack-ansible-rabbitmq_server.git;a=blob;f=tasks/rabbitmq_cluster_create.yml;h=4aaad5adafa42684c0d791b49539069fe1b45635;hb=HEAD#l24

Change-Id: Ia7fc495e315e4a41359641f8be4b3c8bbf7b61fa
Signed-off-by: Fatih Degirmenci <fatih.degirmenci@ericsson.com>