ci: Bump OSA SHA to Pre-Pike
This change bumps osa sha to 9c64c2d8 [1] and includes other changes needed
for it as listed below.

- update ansible role requirements.
- ensure facts are gathered for all the hosts and containers before
running setup-infrastructure.yml playbook.
- stop grepping for failed in the log of setup-hosts.yml playbook due
to existence of rescue block in rabbitmq_cluster_create.yml [2].
- remove the workaround we put for keystone playbook due to not having
the issue anymore.
- reduce the resource needs by updating the user_variables that are
based on upstream osa gate which uses aio. The values for none-aio xci
flavors will be adjusted in a separate change once the sha bump is complete.

[1] https://review.openstack.org/gitweb?p=openstack/openstack-ansible.git;a=commit;h=9c64c2d8bf91a65ca716c1098dc4867eb1cc983a
[2] https://review.openstack.org/gitweb?p=openstack/openstack-ansible-rabbitmq_server.git;a=blob;f=tasks/rabbitmq_cluster_create.yml;h=4aaad5adafa42684c0d791b49539069fe1b45635;hb=HEAD#l24

Change-Id: Ia7fc495e315e4a41359641f8be4b3c8bbf7b61fa
Signed-off-by: Fatih Degirmenci <fatih.degirmenci@ericsson.com>