keepalived: Bring in the fix for iteritems for python3
This change moves the keepalived role version to commit b13e08 in
order to bring the fix applied in upstream, switching from iteritems
to items for python3. [1]

The flavor ha is enabled back due to incorporating the fix to xci and
the note in README.rst is removed as well.

[1] https://github.com/evrardjp/ansible-keepalived/commit/2553ddffd948f3e8e8196cd017d27716b396554a

Change-Id: I6e03f579f1107475f4b4adf54699fc4f924beec3
Signed-off-by: Fatih Degirmenci <fatih.degirmenci@ericsson.com>