keepalived: Bring in the fix for iteritems for python3
This change moves the keepalived role version to commit b13e08 in
order to bring the fix applied in upstream, switching from iteritems
to items for python3. [1]

The flavor ha is enabled back due to incorporating the fix to xci and
the note in README.rst is removed as well.

https://github.com/evrardjp/ansible-keepalived/commit/b13e0840b09154a6d2470f71fea8eaa968525c5b

Change-Id: I6e03f579f1107475f4b4adf54699fc4f924beec3
Signed-off-by: Fatih Degirmenci <fatih.degirmenci@ericsson.com>