bugfix: Add interface files for additional hosts
This is a temporary fix to get noha and ha flavors working.
We should normally use bifrost inventory and have templates for
opnfv, controller, and compute groups rather than per host basis.
But since that is still pending to be done, this change tries to
make sure what we have now continues working.

Change-Id: I9132c474754f4d11d1b9e4b1288b449703994db8
Signed-off-by: Fatih Degirmenci <fatih.degirmenci@ericsson.com>