Disable strict host key checking for galera cluster check
The command to check the galera cluster fails due to having strict
host key checking enabled.

This change disables strict host key checking for galera_container
to ensure we can verify if the database cluster is operational.

OSA does this for aio on their gate and put host key checking
configuration to group_vars which we might need to do as well.

Change-Id: Ifc59581d979c71f1db1b87126659470dc85cc2db
Signed-off-by: Fatih Degirmenci <fatih.degirmenci@ericsson.com>