Create a role which prepares xci for functest
Functest requires xci deployment to have:

- A public neutron network
- A subnet attached to the public neutron network
- The gateway from the subnet attached to one interface

The network and subnet creation are done via a script
instead of using Ansible OpenStack module in order to
get things working asap. So this is yet another quick fix to ensure we 
can move on to functest parts.

Once the release is out, the use of Ansible OpenStack module should be 
investigated and this role should switch to using it.

This patch depends on a pending patch:

https://gerrit.opnfv.org/gerrit/#/c/39517/2

Change-Id: I0635227153087fd9e25b76f0859f82bf48bfb930
Signed-off-by: Manuel Buil <mbuil@suse.com>
Signed-off-by: Fatih Degirmenci <fatih.degirmenci@ericsson.com>