Ensure the pinned versions of components match to the pinned OSA
The versions for the OpenStack services and global requirements
need to be pinned alongside with the OSA SHA1 and the role versions
to ensure we pinned everything.

This change is required for osa-periodic jobs as well since that
job will use sources-branch-updater.sh script and the script updates
these files to ensure things are pinned correctly.

    modified:   ansible-role-requirements.yml
    modified:   global-requirement-pins.txt
    modified:   playbooks/defaults/repo_packages/openstack_services.yml
    modified:   releasenotes/notes/glance-init-config-overrides-d1c8c3dcc50c109a.yaml
    modified:   releasenotes/notes/neutron-init-config-overrides-9d1d2b3b908705ed.yaml
    modified:   releasenotes/notes/trove-init-config-overrides-a78ed428a32adef8.yaml

By doing this change, we do not need to have any
magic to capture updates to those files. Just updating the files
we keep in xci/file folder will be sufficient and they get copied
over during the playbook execution.

Change-Id: Iae0db22574a0368e896132469a8587d1457ce177
Signed-off-by: Fatih Degirmenci <fatih.degirmenci@ericsson.com>