Ensure the pinned versions of components match to the pinned OSA
The versions for the OpenStack services, testing, nova consoles
gnocchi and OpenDaylight need to be pinned alongside with the OSA
SHA1 and the role versions to ensure we pinned everything.

This change is required for osa-periodic jobs as well since that
job will use sources-branch-updater.sh script and have all these
files updated. By doing this change, we do not need to have any
magic to capture updates to those files. Just updating the files
we keep in xci/file folder will be sufficient and they get copied
over during the playbook execution.

Change-Id: Iae0db22574a0368e896132469a8587d1457ce177
Signed-off-by: Fatih Degirmenci <fatih.degirmenci@ericsson.com>