Check if /etc/ssl/certs exists before creating it
In CentOS and Suse, /etc/ssl/certs is a symbolic link to
/etc/ssl/pki/tls/certs.  The Ansible module "file" will fail if it is asked to
create a directory which is already a symbolic link.

This patch will check if /etc/ssl/certs exists before trying to create it.

The same check is done both on the host and the opnfv guest VM.

Also, /etc/certs is only writable by root, so we need a "become: true" clause
to be able to modify it in localhost (but not in opnvf VM).

Change-Id: Iab6c3c162548f84ad6082829e4a7c2ab63d2cfa0
Signed-off-by: Tapio Tallgren <tapio.tallgren@nokia.com>