xci: scripts: Add script to use clean virtual machines for testing
These scripts will prepare a clean OS image based on diskimage-builder
and launch a new VM with it. The purpose of that is to use this virtual
machine as a 'host' to run XCI on it. This way we can ensure that all
tests are being executed on the same clean environment which is ideal
for reproducing and debugging purposes. We also commit the ssh keypair
so we can build it in the image and use it to control the virtual
machine without user interaction. Needless to say that this keypair MUST
NOT be used for anything else apart from the dib build process.

Change-Id: Idec0fc06c82435adc1a4d569b8e578616998de5f
Signed-off-by: Markos Chandras <mchandras@suse.de>