Install docker and shade on opnfv host
Docker is needed for running tests against the deployment.
Shade is needed for managing OpenStack via Ansible.

This change adds tasks to install docker and shade on opnfv
host if it is run as part of CI.

Users should be free to install these if they want so it is
not installed for them by default.

Change-Id: Idfd0f02312cc5e1b0180ed2408755a8c730b987b
Signed-off-by: Fatih Degirmenci <fatih.degirmenci@ericsson.com>