Update patch set 3
Patch Set 3:

> OK but how can we ensure that this is done automatically in the
 > future? 

osa-periodic-jobs will execute source-branch-updater.sh script before it proceeds with the deployment, using updated files. it will then report if the deployment/testing was ok or not and we need to update those files. we can automate that part once we get the jobs up and running. until that happens, we may need to update these files by hand when we bump versions.

Patch-set: 3