Update patch set 3
Patch Set 3:

> But we already have the infrastructure to update both the roles and
 > the services in one go. Anyway, we can pin this now manually and I
 > will look to automate it on another patch

that's what I am saying as well.

the periodic job will update the files for the deployment and tell us if they worked together with the latest sha.

updating the files and sha itself can be automated and we can perhaps make jenkins send a change to gerrit together with shabump, updates files, and so on. this is the part that needs to be done afterwards. and until that happens, we should try to remember to update files together with sha.

Patch-set: 3
Reviewer: Gerrit User 254 <254@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Verified=0