Update patch set 1
Patch Set 1:

> > (1 comment)
 > 
 > Fatih is stopping me :P. We discussed it in IRC today:
 > 
 > <mbuil> fdegir: What do you think about adding the call to the
 > functest role from the scenario? Maybe that way is easier
 > <fdegir> mbuil: I am not so sure about that
 > <fdegir> testing should be separate and the only thing that should
 > change between scenarios is the scenario name or test suites we run

but scenarios should call all the necessary dependencies they need. If that's not the case, then having the prepare-functest as a role instead of a playbook is a mistake and it should be converted to a playbook since we are not treating it as a role. Because right now, having a playbook just for calling a role is pointless.

Patch-set: 1