Add documentation for user guide and sandbox
This change adds the user guide and sandbox information.
The documentation is created by using the information we
have on Wiki and README.rst.

Wiki pages and README.rst will be removed once all the
documentation is moved to docs.opnfv.org.

Parts of the README.rst such as more advanced ways to use
the sandbox will be placed in developer guide and it is
accessible by a link placed in user guide.

Change-Id: I7214a53d7ab39125d2164c5b3468b1ba18933b31
Signed-off-by: Fatih Degirmenci <fatih.degirmenci@ericsson.com>