Update patch set 3
Patch Set 3: Code-Review-1

So we are getting into a bad habit here of adding a new task for every single package that we want to install. This is both ugly and consumes extra time for no reason. How about the following

- add all the packages and wheels to 2 lists in var/opnfv.yaml
- pass that list to the package and pip module

For example

opnfv_required_packages:
  - foo
  - bar

opnfv_required_pip:
  - zz
  - aa

and then you pass the {{ opnfv_required_XXXXX }} to the appropriate module in the playbook.

What do you think?

Patch-set: 3
Label: Code-Review=-1