Update patch set 4
Patch Set 4:

> > > So we are getting into a bad habit here of adding a new task
 > for
 > > > every single package that we want to install. This is both ugly
 > > and
 > > > consumes extra time for no reason. How about the following
 > > >
 > > > - add all the packages and wheels to 2 lists in var/opnfv.yaml
 > > > - pass that list to the package and pip module
 > > >
 > > > For example
 > > >
 > > > opnfv_required_packages:
 > > > - foo
 > > > - bar
 > > >
 > > > opnfv_required_pip:
 > > > - zz
 > > > - aa
 > > >
 > > > and then you pass the {{ opnfv_required_XXXXX }} to the
 > > appropriate
 > > > module in the playbook.
 > > >
 > > > What do you think?
 > >
 > > Sounds like a great idea. I'll do that
 > 
 > The problem is that we might need to create two lists because one
 > is installed always whereas the others are installed only if
 > JENKINS_HOME  exists

What is the reason to install packages in CI and not in local deployments? Why do we want to make such a distinction? In an ideal world the CI should test what developers are deploying on their systems. But if there are valid reasons for CI to behave differently then we need to restructure the entire layout to make it more flexible so we may have to use your original approach for now.

Patch-set: 4
Label: Verified=0