xci: scripts: start-new-vm.sh: Do not destroy user's configuration file
Since this script may be used by developers, we shouldn't destroy
the existing ssh configuration file but rather create a new one
just for XCI. This also move the destruction of previous instances
sooner, even before building the OS image so we can free resources
as soon as possible. Finally, this also uses the hostname to ssh to the
new VM instead of its IP to ensure that everything is setup up properly
on the physical host.

Change-Id: I185befd8943fdb8ebcdb635aefb4d4b03473258b
Signed-off-by: Markos Chandras <mchandras@suse.de>