xci: Deploy XCI in a clean virtual machine
Deploying XCI on developer's host causes far too many problems for
numerous reasons. For examples, the developer's machine could be
semi-broken, or conflicting packages could be installed that prevent
certain XCI components to work. Finally, the destruction script pretends
to be an uninstaller but it always leaves artifacts behind as XCI
evolves. The truth is that we don't need to run XCI natively on the host
since we do almost everything inside VMs. As such, we may as well just
deploy XCI inside a VM similar to what Jenkins does already. The benefit
from that is that we always get reproducible results, since the VM is
always in a clean state. Moreover, developers can manipulate the
installation using regular VM features like snapshotting etc. The
drawback is that for nested virtulization to work efficiently we raise
the hardware requirements but that shouldn't be much of an issue given
that XCI already needs a lot of hardware resources already. This patch
renames xci-deploy.sh to xci-deploy-core.sh and creates a new
xci-deploy.sh which is a wrapper around that script which will prepare
the new VM for XCI deployment.

Change-Id: I4f502b553b840093a1adbb9ebb15059c4011359f
Signed-off-by: Markos Chandras <mchandras@suse.de>