Update patch set 3
Patch Set 3: Code-Review-2

Needs thinking because the hooks can be abused by scenarios to implement everything in bash. The original goal here was for scenarios to have a way to override variables (for example, VM specs) so maybe the hooks need to be dropped and only source the file only for variables.

Patch-set: 3
Label: Code-Review=-2