Update patch set 1
Patch Set 1:

Or maybe having a configuration file in the xci/ directory to include this additional path for roles may not be too bad I don't know. Also this patch assumes that scenarios and role as the same thing (ie the root of scenario is the role) but that's not true for projects which simply embed the role somewhere deeper in the project directory. We need to think of a common solution for both cases.

Patch-set: 1