Update patch set 12
Patch Set 12: Code-Review-1

So this patchset doesn't break our default case which is good but can you also do revision of this patchset that enables Ceph by default so we can see whether this actually works or not? If that passes on Ubuntu (I know it will fail on SUSE) then you can revert it to this version so we can review it.

Patch-set: 12
Label: Code-Review=-1