SDF prototype for xci-baremetal input
This prototype is based on a template system that shall avoid divergences of a
same configuration for all sdf files. By that way, a same configuration will
evolve in all sdf at the same time.

Installer have to agree on the same convention as much as possible, but specific
configuration is possible by this sytem.

Change-Id: I84d43fedfda4fc1dbf50e7285ddf7d3dc3781f96
Signed-off-by: Blaisonneau David <david.blaisonneau@orange.com>