Update patch set 5
Patch Set 5:

Thanks a lot for this David, looks pretty good.

I will have more comments, especially regarding the k8s/openstack roles as I believe we need to think these as part of OPNFV scenarios and move them to their final places which we can discuss more later. So please expect more comments over time while we work on the other things.

The thing we currently focus is how to ensure we can split what we do as part of bifrost installation, node creation/provisioning/configuration which is what you already do as part of this patch. We can perhaps take those parts out of this patch and start getting them in for real by incorporating virtual deployments as well. 

Another thing that got our attention is the need to further isolate ourselves from the physical host where the virtual deployments are done and from the jumphost where the baremetal deployments are driven.

We intend to take the opnfv vm out of bifrost's control and push everything related to bifrost into the vm rather than doing it on physical host - which is the root of most of the problems reported by users. Again, this part will need to be done as part of the initial work.

I hope it is fine for you that I take some of the roles you have in this change and send few smaller changes to show what I'm thinking. We can then work on these changes together.

Further information is on https://etherpad.opnfv.org/p/xci-structure if anyone wants to know more.

Patch-set: 5
Reviewer: Gerrit User 254 <254@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Verified=0