Update patch set 2
Patch Set 2:

As we discussed on irc and as you highlight in your comment, we need to have a common way for handling the scenarios. To me whether a scenarios is generic or specific does not matter at this phase so we need common way of handling all the scenarios.

My thinking when I proposed this change and tried to summarize irc is based on how osa advices people to create new roles. If we apply that to OPNFV scenarios, I end up with treating scenarios as roles, thus proposing this patch.

When it comes to how the roles for generic scenarios get connected to XCI; same as the roles for the specific scenarios which you proposed in change; https://gerrit.opnfv.org/gerrit/#/c/44881/ and put in to file opnfv-scenario-requirements.yml.

Patch-set: 2