Update patch set 3
Patch Set 3: Code-Review-1

I don't know what should we do with this change tbh...

As mentioned earlier, this will change what the users find on their machines upon completion of the deployment; single VM with all the nodes in it which I'm not too fond of due to

- possibility of us going back to original way when we fix the things as documented on etherpad which will be yet another change, probably confusing users.
- diverging from how the baremetal deployment looks. the way we have opnfv and target nodes created next to each other is on purpose and try to be in similar look.

Also, if we think about other initiatives we have, this is not future proof, especially for LaaS or any machine which is already a VM. When this is run on one of this type of machines, the level of nesting will be 3 for the actual nodes. The VMs created on OpenStack will be on 4th level. (the host, the opnfv clean vm, the openstack nodes, and finally the instances created on openstack are all VMs.)

Patch-set: 3
Reviewer: Gerrit User 254 <254@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Code-Review=-1